arm/arm64: Remove dcache_line_size global variable
Compute the dcache line size when doing dcache maintenance instead of using
a global variable computed in setup(), which allows us to do dcache
maintenance at any point in the boot process. This will be useful for
running as an EFI app it also makes it possible to align our implementation
with the Linux kernel.

For consistency, the arm code has been similary modified.