Revert "arm/arm64: mmu_disable: Clean and invalidate before disabling"
This reverts commit fc58684bc47b7d07d75098fdfddb6083e9b12104.

Commit 410b3bf09e76 ("arm/arm64: Perform dcache clean + invalidate after
turning MMU off") is perfectly fine as per the architecture. The problem
that commit fc58684bc47b ("arm/arm64: mmu_disable: Clean and invalidate
before disabling") is trying to solve is caused by the fact that
__phys_offset and __phys_end are computed with the MMU on and still in the
cache, but they are read with the MMU off from main memory, which results
in using the initial values of 0 for the variables.

We will fix all the cache-related problems in one fell swoop when we will
deactivate the MMU in assembly code before branching to setup().